services/ingestion-pipeline/requirements.txt
kafka-python==2.0.2
qdrant-client==1.9.1
openai==1.30.1
psycopg[binary]==3.1.19
python-dotenv==1.0.1
sqlalchemy==2.0.30
